refactor: CacheConfig infrastructure 레이어로 이동 및 캐시 설정 및 빌더로 중복 제거
