+++ backend/src/main/kotlin/com/festago/common/cache/CacheManageEventListener.kt
@@ -1,7 +1,9 @@
 package com.festago.common.cache
 
 import io.github.oshai.kotlinlogging.KotlinLogging
+import jakarta.annotation.PreDestroy
 import org.springframework.cache.CacheManager
+import org.springframework.cache.caffeine.CaffeineCache
 import org.springframework.context.annotation.Profile
 import org.springframework.context.event.EventListener
 import org.springframework.scheduling.annotation.Async
@@ -31,9 +33,20 @@ internal class CacheManageEventListener(
         val cache = cacheManager.getCache(cacheName) ?: return
         cache.evict(event.key)
     }
+
+    @PreDestroy
+    protected fun logCacheStats() {
+        for (cacheName in cacheManager.cacheNames) {
+            val cache = cacheManager.getCache(cacheName)
+            if (cache is CaffeineCache) {
+                log.info {
+                    "CacheName=${cacheName} CacheStats=${cache.nativeCache.stats()}"
+                }
+            }
+        }
+    }
 }
 
 data class CacheInvalidateCommandEvent(val cacheName: String)
 
 data class CacheEvictCommandEvent(val cacheName: String, val key: String)
-
