+++ backend/src/main/kotlin/com/festago/config/TransactionalConfig.kt
@@ -1,0 +1,8 @@
+package com.festago.config
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.transaction.annotation.EnableTransactionManagement
+
+@Configuration
+@EnableTransactionManagement(order = 0)
+class TransactionalConfig
