+++ backend/src/main/kotlin/com/festago/festival/application/query/FestivalV1QueryService.kt
@@ -1,31 +1,15 @@
 package com.festago.festival.application.query
 
-import com.festago.common.cache.CacheInvalidateCommandEvent
 import com.festago.common.exception.ValidException
 import com.festago.festival.dto.FestivalV1QueryRequest
 import com.festago.festival.dto.FestivalV1Response
-import com.festago.festival.dto.event.FestivalCreatedEvent
-import com.festago.festival.dto.event.FestivalDeletedEvent
-import com.festago.festival.dto.event.FestivalUpdatedEvent
 import com.festago.festival.infrastructure.repository.query.FestivalSearchCondition
 import com.festago.festival.infrastructure.repository.query.FestivalV1QueryDslRepository
-import com.festago.stage.dto.event.StageCreatedEvent
-import com.festago.stage.dto.event.StageDeletedEvent
-import com.festago.stage.dto.event.StageUpdatedEvent
-import com.github.benmanes.caffeine.cache.Caffeine
 import java.time.Clock
-import java.time.Duration
 import java.time.LocalDate
-import org.springframework.cache.Cache
 import org.springframework.cache.annotation.Cacheable
-import org.springframework.cache.caffeine.CaffeineCache
-import org.springframework.context.ApplicationEventPublisher
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.context.event.EventListener
 import org.springframework.data.domain.Pageable
 import org.springframework.data.domain.Slice
-import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
@@ -69,35 +53,3 @@ class FestivalV1QueryService(
         const val CACHE_NAME: String = "FIND_FESTIVALS_V1"
     }
 }
-
-@Configuration
-class FestivalV1QueryServiceCacheConfig(
-    private val eventPublisher: ApplicationEventPublisher,
-) {
-
-    @Bean
-    fun festivalV1QueryServiceCache(): Cache {
-        return CaffeineCache(
-            FestivalV1QueryService.CACHE_NAME, Caffeine.newBuilder()
-                .expireAfterWrite(Duration.ofHours(1))
-                .maximumSize(10)
-                .build()
-        )
-    }
-
-    @EventListener(value = [FestivalCreatedEvent::class, FestivalUpdatedEvent::class, FestivalDeletedEvent::class])
-    fun invalidateCacheWhenFestivalModified() {
-        eventPublisher.publishEvent(CacheInvalidateCommandEvent(FestivalV1QueryService.CACHE_NAME))
-    }
-
-    @EventListener(value = [StageCreatedEvent::class, StageUpdatedEvent::class, StageDeletedEvent::class])
-    fun invalidateCacheWhenStageModified() {
-        eventPublisher.publishEvent(CacheInvalidateCommandEvent(FestivalV1QueryService.CACHE_NAME))
-    }
-
-    @Scheduled(cron = "0 0 0 * * *")
-    fun invalidateCacheAtMidnight() {
-        eventPublisher.publishEvent(CacheInvalidateCommandEvent(FestivalV1QueryService.CACHE_NAME))
-    }
-}
-
